Move transaction helper function to helpers folder rather than util
